requirements.txt
opencv-python
vietocr
torchmetrics
tensorboard
streamlit
Pillow
numpy
pyyaml
tqdm
matplotlib